くそネームスペースの改築。
