--- PracticeChat/PracticeChat/ATH/FontManager.h
@@ -1,20 +1,0 @@
-#pragma once
-
-namespace ATH {
-	// 0 = ノーマルテキスト
-	// 1 = ボールドテキスト
-	// 2 = イタリックテキスト
-	// 3 = ボールドイタリック
-
-	class FontManager
-	{
-		void _makeFont(int kind, int fontsize);
-	public:
-		siv::Font f[4];
-		void setFontsize(int kind, int FontSize);
-		FontManager();
-		~FontManager();
-	};
-
-
-}
