--- PracticeChat/PracticeChat/FontExperiment.h
@@ -1,10 +1,0 @@
-#pragma once
-class FontExperiment
-{
-public:
-	FontExperiment();
-	~FontExperiment();
-
-	int MainLoop();
-};
-
