+++ PracticeChat/PracticeChat/PracticeChat.vcxproj
@@ -167,7 +167,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="ATH\FontManager.cpp" />
+    <ClCompile Include="Experiment\ATH\FontManager.cpp" />
     <ClCompile Include="Main.cpp" />
     <ClCompile Include="tcpframework\test.cpp" />
   </ItemGroup>
@@ -292,7 +292,7 @@
     <ClInclude Include="Experiment\TestElipmoc.hpp" />
     <ClInclude Include="Experiment\Test_A_TO_HA.hpp" />
     <ClInclude Include="Experiment\Test☆彡.hpp" />
-    <ClInclude Include="ATH\FontManager.h" />
+    <ClInclude Include="Experiment\ATH\FontManager.h" />
     <ClInclude Include="tcpframework\test.hpp" />
     <ClInclude Include="UseSiv3D.hpp" />
   </ItemGroup>
