+++ PracticeChat/PracticeChat/PracticeChat.vcxproj.filters
@@ -84,7 +84,7 @@
     <ClCompile Include="tcpframework\test.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="ATH\FontManager.cpp">
+    <ClCompile Include="Experiment\ATH\FontManager.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
   </ItemGroup>
@@ -431,7 +431,7 @@
     <ClInclude Include="Experiment\Test☆彡.hpp">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
-    <ClInclude Include="ATH\FontManager.h">
+    <ClInclude Include="Experiment\ATH\FontManager.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
   </ItemGroup>
